LayoutTests/accessibility/aria-hidden.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../fast/js/resources/js-test-style.css">
<script>
var successfullyParsed = false;
</script>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">
<h1 aria-hidden="true">h1 <b>test</b></h1>
<h2 aria-hidden="false">h2</h2>
<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the aria-hidden attribute works correctly with accessibility. The H1 element (and its children) should not appear in the AX hierarchy. The H2 element should be the first child");

    if (window.accessibilityController) {

          var body = document.getElementById("body");
          body.focus();
          var h2 = accessibilityController.focusedElement.childAtIndex(0);
          var pattern = "AXTitle: h2";
          var succeeded = h2.allAttributes().indexOf(pattern) != -1;
          shouldBe("succeeded", "true");
    }

    successfullyParsed = true;
</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
